Create BSPD/BSPD-revamp/BSPD-revamp-layout
Trying to get BSPD a little smaller for fitting into the GLV box. Somewhat worked. Still a few problems with  how the front and back ground zones connect, and the differences between the GNDPWR pin and POWER pin (intentional? but still a problem) causing two different disconnected grounds on the PCB layout.